+++ luqum/naming.py
@@ -5,6 +5,7 @@ and retrieve their positions in the query text.
 
 This module adds support for that.
 """
+from . import tree
 from .visitor import TreeVisitor
 
 
@@ -23,82 +24,58 @@ def get_name(node):
 class TreeAutoNamer(TreeVisitor):
     # helper for :py:func:`tree_name_index`
 
-    def __init__(self):
+    DEFAULT_TARGETS = (tree.Range, tree.Term, tree.Fuzzy, tree.BaseApprox)
+    """by default targets are the one translated to a leaf term in elasticsearch
+
+    :param tuple targets: class of elements that should receive a name
+    :param bool all_names: shall we name children of named elements ?
+    """
+
+    LETTERS = "abcdefghilklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+    _letter_pos = {i: l for i, l in enumerate(LETTERS)}
+    _pos_letter = {l: i for i, l in enumerate(LETTERS)}
+
+    def __init__(self, targets=None, all_names=False):
+        self.targets = targets if targets is not None else self.DEFAULT_TARGETS
+        self.all_names = all_names
         super().__init__(track_parents=False)
 
-    def visit_base_operation(self, node, context):
-        # operations are the splitting point, we name them and make children subnames
-        names = context["names"]
-        set_name(node, ("_".join(names)))
-        for i, c in enumerate(node.children):
-            yield from self.visit_iter(c, context={"names": names + [str(i)]})
-
-    def visit_term(self, node, context=None):
-        names = context["names"]
-        set_name(node, ("_".join(names)))
-        return []
-
-    def visit_range(self, node, context=None):
-        names = context["names"]
-        set_name(node, ("_".join(names)))
-        # no need to visit children
-        return []
+    def generic_visit(self, node, context):
+        visit_children = True
+        if isinstance(node, self.targets):
+            # add name
+            self.context["name"] = name = self.next_name(context["name"])
+            set_name(node, name)
+            # track name / path correspondance
+            context["name_to_path"][name] = tuple(context["path"])
+            visit_children = self.all_names
+        if visit_children:
+            # continue with path trackings
+            for i, child in enumerate(node.children):
+                child_context = dict(context, path=context["path"] + [i])
+                yield from self.visit_iter(child, context=child_context)
+        else:
+            yield node
 
     def visit(self, node):
-        return list(self.visit_iter(node, context={"names": ["0"]}))
+        """visit the tree and add names to nodes while tracking their path
+        """
+        context = {"path": [], "name": None, "name_to_path": {}}
+        list(self.visit_iter(node, context=context))
+        return context["name_to_path"]
 
 
-def auto_name(tree):
+def auto_name(tree, targets=None, all_names=False):
     """Automatically add names to nodes of a parse tree.
 
-    We add them to terminal nodes : range, phrases and words, as this is where it is useful,
-    but also on operations, to easily grab the group.
+    We add them to terminal nodes : range, phrases and words, as this is where it is useful.
+
+    :return: a dict giving the path to a the children having this name
     """
-    TreeAutoNamer().visit(tree)
+    return TreeAutoNamer().visit(tree, targets, all_names)
 
 
-class NameIndexer(TreeVisitor):
-    # helper for :py:func:`tree_name_index`
-
-    def __init__(self):
-        super().__init__(track_parents=True)
-
-    def generic_visit(self, node, context):
-        # visit children
-        sub_names = list(super().generic_visit(node, context))
-        name = get_name(node)
-        root_node = not context.get("parents")
-        if name is not None or root_node:
-            str_repr = str(node)
-            # search for subnodes position
-            subnodes_pos = []
-            idx = 0
-            for (subname, sub_repr, sub_subnodes_pos) in sub_names:
-                pos = str_repr.find(sub_repr, idx)
-                if pos >= 0:  # pragma: no branch
-                    length = len(sub_repr)
-                    subnodes_pos.append((subname, pos, length, sub_subnodes_pos))
-                    idx = pos + length
-            sub_names = [(name, str_repr, subnodes_pos)]
-        yield from sub_names
-
-    def __call__(self, node):
-        subnames = self.visit(node)
-        # by construction, root node, only return one entry
-        name, str_repr, subnodes_pos = subnames[0]
-        if name is not None:
-            # resolve last level
-            subnodes_pos = [(name, 0, len(str_repr), subnodes_pos)]
-        return subnodes_pos
-
-
-def _flatten_name_index(subnodes_pos, start_pos=0):
-    for name, pos, length, children in subnodes_pos:
-        yield name, start_pos + pos, length
-        yield from _flatten_name_index(children, start_pos + pos)
-
-
-def name_index(tree):
+def name_index(tree, name_to_path):
     """Given a tree with names, give the index of each group in the string representation.
     also gives the node type.
 
@@ -111,10 +88,6 @@ def name_index(tree):
     :param tree: a luqum parse tree
     :return dict: mapping each name to a `(start position, length)` tuple
     """
-    subnodes_pos = NameIndexer()(tree)
-    # flatten the hierarchy
-    result = {name: (pos, length) for name, pos, length in _flatten_name_index(subnodes_pos)}
-    return result
 
 
 def extract(expr, name, name_index):
