+++ setup.py
@@ -35,5 +35,4 @@ setup(
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
     ],
-    test_suite='luqum.tests'
 )
